Reorganizing and adding new tests folder
